third_party/WebKit/LayoutTests/platform/mac-mac10.11/external/wpt/XMLHttpRequest/preserve-ua-header-on-redirect-expected.txt

This is a testharness.js-based test.
FAIL XMLHttpRequest: User-Agent header is preserved on redirect assert_equals: expected "User-Agent: Mozilla/5.0 (Macintosh; Intel Mac OS X 10_11_3) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/99.77.34.5 Safari/537.36\n" but got "{\"error\": {\"message\": \"Traceback (most recent call last):\\n  File \\"/b/c/b/mac_layout/src/third_party/WebKit/Tools/Scripts/webkitpy/thirdparty/wpt/wpt/tools/wptserve/wptserve/handlers.py\\", line 246, in __call__\n    rv = self.func(request, response)\n  File \\"/b/c/b/mac_layout/src/third_party/WebKit/LayoutTests/external/wpt/XMLHttpRequest/resources/inspect-headers.py\\", line 31, in main\n    result = get_response(request.raw_headers, filter_value, filter_name)\nAttributeError: 'Request' object has no attribute 'raw_headers'\n\", \"code\": 500}}"
FAIL XMLHttpRequest: User-Agent header is preserved on redirect 1 assert_equals: expected "User-Agent: TEST\n" but got "{\"error\": {\"message\": \"Traceback (most recent call last):\\n  File \\"/b/c/b/mac_layout/src/third_party/WebKit/Tools/Scripts/webkitpy/thirdparty/wpt/wpt/tools/wptserve/wptserve/handlers.py\\", line 246, in __call__\n    rv = self.func(request, response)\n  File \\"/b/c/b/mac_layout/src/third_party/WebKit/LayoutTests/external/wpt/XMLHttpRequest/resources/inspect-headers.py\\", line 31, in main\n    result = get_response(request.raw_headers, filter_value, filter_name)\nAttributeError: 'Request' object has no attribute 'raw_headers'\n\", \"code\": 500}}"
Harness: the test ran to completion.
